now use the Point(int, int) constructor
git-svn-id: http://svn.code.sf.net/p/graphiti-editor/code@565 04dbf5ff-8551-0410-9f1f-ce2c5325ca72

+++ plugins/net.sf.graphiti.ui/src/net/sf/graphiti/ui/figure/PolygonPortAnchor.java
@@ -104,7 +104,7 @@ public class PolygonPortAnchor extends AbstractConnectionAnchor {
 				float d = start.y - p * start.x;
 
 				// Compute xAnchor
-				float xAnchor;
+				int xAnchor;
 				if (run == 0) {
 					// Line equation: x = center.x
 					xAnchor = center.x;
@@ -114,11 +114,11 @@ public class PolygonPortAnchor extends AbstractConnectionAnchor {
 					float rise = (reference.y - center.y);
 					float a = rise / run;
 					float b = center.y - a * center.x;
-					xAnchor = (d - b) / (a - p);
+					xAnchor = (int) ((d - b) / (a - p));
 				}
 
 				// yAnchor is just y = px + d
-				float yAnchor = p * xAnchor + d;
+				int yAnchor = (int) (p * xAnchor + d);
 				return new Point(xAnchor, yAnchor);
 			}
 		}
